Visual Studio 2008/RestApiSDK/PayPal/Api/Payments/AmountDetails.cs
using System;
using Newtonsoft.Json;
using Newtonsoft.Json.Serialization;
using PayPal.Api.Payments;

namespace PayPal.Api.Payments
{

	/// <summary>
	/// 
    /// </summary>
	public class AmountDetails : Resource  
	{

		/// <summary>
		/// subtotal
    	/// </summary>
		[JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
		public string subtotal
		{
			get;
			set;
		}
		

		/// <summary>
		/// tax
    	/// </summary>
		[JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
		public string tax
		{
			get;
			set;
		}
		

		/// <summary>
		/// shipping
    	/// </summary>
		[JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
		public string shipping
		{
			get;
			set;
		}
		

		/// <summary>
		/// fee
    	/// </summary>
		[JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
		public string fee
		{
			get;
			set;
		}
		

		/// <summary>
		/// Converts the object to JSON string
		/// </summary>
		public new string ConvertToJson() 
    	{ 
    		return JsonFormatter.ConvertToJson(this);
    	}
    	
	}
}
